refactor: Rename landmark package to place for improved clarity
