+++ pida-core/core-domain/src/main/kotlin/com/pida/place/Landmark.kt
@@ -1,4 +1,4 @@
-package com.pida.landmark
+package com.pida.place
 
 import com.pida.support.geo.GeoJson
 import com.pida.support.geo.Region
